Revert change of production environment context
Although the provided address is almost certainly not going to be our final production address as it's on localhost, I don't think we have any reason to change the production context at this time. We should save any change until we have a clearer picture of our production environment - ultimately this should likely be decided by the DevOps team.