removed button from skillType
There was a button for save and close but two way binding was implemented so auto save is auto matically implied by removing the save button and just having the close button. Before when using the close button it will commit changes to the front end but not persist change to the backend. Now the user only has the option to close which now implies there is no way to revert changes unless they change it back.