Can't activate or deactivate a category on a bucket without the category on it.
From our understanding of this, we are adding and removing buckets activity state. So if the bucket is inactive we can toggle it to active from the category view. LIkewise we can do the reverse. But this requires a bucket to have a categoryId == our category which cannot exist if the category is just now being created. Therefore we should just allow name and possibly skilltype to be added to each category as we create it.