For delete to actually work, we need to trigger on Observable.complete()
Our backend does not send any json in the body of the DELETE response so the observable does not actually file a success. The best work around I can come up with at this time is to just reload the categories listing after the request completes regardless of success or failure.